AA: Move error codes from AA package to base package (#654)
* changes

* changes

* changes

* changes

* changes

* changes
