+++ packages/client/base/src/index.ts
@@ -2,3 +2,4 @@ export * from "./utils";
 export * from "./types";
 export * from "./services";
 export * from "./consts";
+export * from "./error";
