+++ packages/client/wallets/smart-wallet/package.json
@@ -29,6 +29,7 @@
         "unlink:common-sdk-base": "yalc remove @crossmint/common-sdk-base && pnpm i"
     },
     "dependencies": {
+        "@crossmint/client-sdk-base": "workspace:*",
         "@crossmint/common-sdk-base": "workspace:*",
         "@datadog/browser-logs": "4.42.2",
         "@zerodev/ecdsa-validator": "5.3.1",
