+++ packages/client/wallets/smart-wallet/src/api/BaseCrossmintService.ts
@@ -1,6 +1,6 @@
+import { CrossmintServiceError } from "@crossmint/client-sdk-base";
 import { validateAPIKey } from "@crossmint/common-sdk-base";
 
-import { CrossmintServiceError } from "../error";
 import { CROSSMINT_DEV_URL, CROSSMINT_PROD_URL, CROSSMINT_STG_URL } from "../utils/constants";
 import { logPerformance } from "../utils/log";
 import { APIErrorService } from "./APIErrorService";
