+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/EVMSmartWallet.ts
@@ -1,8 +1,9 @@
 import { logError, logInfo } from "@/services/logging";
 import { type HttpTransport, type PublicClient, isAddress, publicActions } from "viem";
 
+import { TransferError } from "@crossmint/client-sdk-base";
+
 import type { CrossmintWalletService } from "../../api/CrossmintWalletService";
-import { TransferError } from "../../error";
 import type { TransferType } from "../../types/Tokens";
 import { SmartWalletClient } from "../../types/internal";
 import { SCW_SERVICE } from "../../utils/constants";
