+++ packages/client/wallets/smart-wallet/src/error/index.ts
@@ -1,50 +1,10 @@
 import { PasskeyDisplay, SignerDisplay } from "@/types/API";
 
-export const SmartWalletErrorCode = {
-    NOT_AUTHORIZED: "not-authorized",
-    CROSSMINT_SERVICE: "crossmint-service",
-    JWT_EXPIRED: "not-authorized.jwt-expired",
-    JWT_INVALID: "not-authorized.jwt-invalid",
-    JWT_DECRYPTION: "not-authorized.jwt-decryption",
-    JWT_IDENTIFIER: "not-authorized.jwt-identifier",
-    OUT_OF_CREDITS: "out-of-credits",
-    TRANSFER: "smart-wallet:transfer",
-    SMART_WALLETS_NOT_ENABLED: "smart-wallet:not-enabled",
-    USER_WALLET_ALREADY_CREATED: "smart-wallet:user-wallet-already-created",
-    WALLET_CONFIG: "smart-wallet:config",
-    ADMIN_MISMATCH: "smart-wallet:config.admin-mismatch",
-    PASSKEY_MISMATCH: "smart-wallet:config.passkey-mismatch",
-    ADMIN_SIGNER_ALREADY_USED: "smart-wallet:config.admin-signer-already-used",
-    PASSKEY_PROMPT: "smart-wallet:passkey.prompt",
-    PASSKEY_INCOMPATIBLE_AUTHENTICATOR: "smart-wallet:passkey.incompatible-authenticator",
-    PASSKEY_REGISTRATION: "smart-wallet:passkey.registration",
-    UNCATEGORIZED: "smart-wallet:uncategorized", // smart wallet specific catch-all error code
-} as const;
-export type SmartWalletErrorCode = (typeof SmartWalletErrorCode)[keyof typeof SmartWalletErrorCode];
+import { CrossmintSDKError, SmartWalletErrorCode } from "@crossmint/client-sdk-base";
 
-export class SmartWalletError extends Error {
-    public readonly code: SmartWalletErrorCode;
-    public readonly details?: string;
-
+export class SmartWalletError extends CrossmintSDKError {
     constructor(message: string, details?: string, code: SmartWalletErrorCode = SmartWalletErrorCode.UNCATEGORIZED) {
-        super(message);
-        this.details = details;
-        this.code = code;
-    }
-}
-
-export class TransferError extends SmartWalletError {
-    constructor(message: string) {
-        super(message, undefined, SmartWalletErrorCode.TRANSFER);
-    }
-}
-
-export class CrossmintServiceError extends SmartWalletError {
-    public status?: number;
-
-    constructor(message: string, status?: number) {
-        super(message, undefined, SmartWalletErrorCode.CROSSMINT_SERVICE);
-        this.status = status;
+        super(message, code, details);
     }
 }
 
@@ -67,50 +27,6 @@ export class PasskeyMismatchError extends SmartWalletError {
         super(message, SmartWalletErrorCode.PASSKEY_MISMATCH);
         this.required = required;
         this.used = used;
-    }
-}
-
-export class NotAuthorizedError extends SmartWalletError {
-    constructor(message: string) {
-        super(message, undefined, SmartWalletErrorCode.NOT_AUTHORIZED);
-    }
-}
-
-export class JWTExpiredError extends NotAuthorizedError {
-    public readonly code = SmartWalletErrorCode.JWT_EXPIRED;
-
-    /**
-     * The expiry time of the JWT as an ISO 8601 timestamp.
-     */
-    public readonly expiredAt: string;
-
-    constructor(expiredAt: Date) {
-        super(`JWT provided expired at timestamp ${expiredAt}`);
-        this.expiredAt = expiredAt.toISOString();
-    }
-}
-
-export class JWTInvalidError extends NotAuthorizedError {
-    public readonly code = SmartWalletErrorCode.JWT_INVALID;
-    constructor() {
-        super("Invalid JWT provided");
-    }
-}
-
-export class JWTDecryptionError extends NotAuthorizedError {
-    public readonly code = SmartWalletErrorCode.JWT_DECRYPTION;
-    constructor() {
-        super("Error decrypting JWT");
-    }
-}
-
-export class JWTIdentifierError extends NotAuthorizedError {
-    public readonly code = SmartWalletErrorCode.JWT_IDENTIFIER;
-    public readonly identifierKey: string;
-
-    constructor(identifierKey: string) {
-        super(`Missing required identifier '${identifierKey}' in the JWT`);
-        this.identifierKey = identifierKey;
     }
 }
 
@@ -158,16 +74,6 @@ export class PasskeyIncompatibleAuthenticatorError extends SmartWalletError {
             SmartWalletErrorCode.PASSKEY_INCOMPATIBLE_AUTHENTICATOR
         );
         this.passkeyName = passkeyName;
-    }
-}
-
-export class OutOfCreditsError extends SmartWalletError {
-    constructor(message?: string) {
-        super(
-            "You've run out of Crossmint API credits. Visit https://docs.crossmint.com/docs/errors for more information",
-            undefined,
-            SmartWalletErrorCode.OUT_OF_CREDITS
-        );
     }
 }
